// src/app/types/generated/tgm.Application.TgAccounts.Features.Chating.DTOs.ts
/* THIS (.ts) FILE IS GENERATED BY Tapper */
/* eslint-disable */
/* tslint:disable */
import type { TgServiceState } from './tgm.Domain.TgAccounts.Enums';

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.AccountDTO */
export type AccountDTO = {
    /** Transpiled from System.Guid */
    id: string;
    /** Transpiled from string */
    firstName: string;
    /** Transpiled from string */
    phoneNumber: string;
    /** Transpiled from tgm.Domain.TgAccounts.Enums.TgServiceState */
    serviceState: TgServiceState;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.ChatDTO */
export type ChatDTO = {
    /** Transpiled from long */
    id: number;
    /** Transpiled from string */
    title: string;
    /** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.MessageDTO? */
    lastMessage?: MessageDTO;
    /** Transpiled from System.Collections.Generic.List<tgm.Application.TgAccounts.Features.Chating.DTOs.ChatPositionDTO> */
    positions: ChatPositionDTO[];
    /** Transpiled from bool */
    isOrdered: boolean;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.ChatFolderInfoDTO */
export type ChatFolderInfoDTO = {
    /** Transpiled from int */
    id: number;
    /** Transpiled from string */
    name: string;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.ChatListDTO */
export type ChatListDTO = {
    /** Transpiled from string */
    type: string;
    /** Transpiled from int */
    chatFolderId?: number;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.ChatPositionDTO */
export type ChatPositionDTO = {
    /** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.ChatListDTO */
    list: ChatListDTO;
    /** Transpiled from long */
    order: number;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.MessageDTO */
export type MessageDTO = {
    /** Transpiled from long */
    id: number;
    /** Transpiled from string */
    content: string;
    /** Transpiled from int */
    date: number;
    /** Transpiled from bool */
    isOutgoing: boolean;
    /** Transpiled from bool */
    isPending: boolean;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.UserDTO */
export type UserDTO = {
    /** Transpiled from long */
    id: number;
    /** Transpiled from string */
    firstName: string;
    /** Transpiled from string */
    lastname: string;
    /** Transpiled from string */
    phoneNumber: string;
    /** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.UserStatusDTO */
    status: UserStatusDTO;
    /** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.UserTypeDTO */
    type: UserTypeDTO;
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.UserStatusDTO */
export enum UserStatusDTO {
    Empty = 0,
    LastMonth = 1,
    LastWeek = 2,
    Offline = 3,
    Online = 4,
    Recently = 5,
}

/** Transpiled from tgm.Application.TgAccounts.Features.Chating.DTOs.UserTypeDTO */
export enum UserTypeDTO {
    Bot = 0,
    Deleted = 1,
    Regular = 2,
    Unknown = 3,
}